restructure code with domain-based approach
