+++ http/sajjad/http-client/useraccount/model.go
@@ -1,9 +1,11 @@
-package models
+package useraccount
 
+// Account represents a user account with attributes
 type Account struct {
 	Attributes AccountAttributes `json:"attributes"`
 }
 
+// AccountAttributes contains the account's attributes
 type AccountAttributes struct {
 	ID      int    `json:"id"`
 	UserID  int    `json:"user_id"`
